+++ my-app/src/pages/feed/post/postContentImg.style.js
@@ -1,9 +1,7 @@
 import styled from "styled-components";
 
 export const Contentimg = styled.img`
-    margin-left: 42px;
-    width: calc(100% - 42px);
-    max-width: 600px;
+    width: 100%;
     object-fit: cover;
     border: 0.5px solid #dbdbdb;
     border-radius: 10px;
